+++ app/services/idv/steps/threat_metrix_step_helper.rb
@@ -3,6 +3,7 @@
 module Idv
   module Steps
     module ThreatMetrixStepHelper
+      include ThreatMetrixHelper
       def threatmetrix_view_variables(updating_ssn)
         session_id = generate_threatmetrix_session_id(updating_ssn)
 
@@ -27,37 +28,6 @@ module Idv
         else
           true
         end
-      end
-
-      # @return [Array<String>]
-      def threatmetrix_javascript_urls(session_id)
-        sources = if IdentityConfig.store.lexisnexis_threatmetrix_mock_enabled
-                    Rails.application.config.asset_sources.get_sources('mock-device-profiling')
-                  else
-                    ['https://h.online-metrix.net/fp/tags.js']
-                  end
-
-        sources.map do |source|
-          UriService.add_params(
-            source,
-            org_id: IdentityConfig.store.lexisnexis_threatmetrix_org_id,
-            session_id: session_id,
-          )
-        end
-      end
-
-      def threatmetrix_iframe_url(session_id)
-        source = if IdentityConfig.store.lexisnexis_threatmetrix_mock_enabled
-                   Rails.application.routes.url_helpers.test_device_profiling_iframe_url
-                 else
-                   'https://h.online-metrix.net/fp/tags'
-                 end
-
-        UriService.add_params(
-          source,
-          org_id: IdentityConfig.store.lexisnexis_threatmetrix_org_id,
-          session_id: session_id,
-        )
       end
     end
   end
