+++ lib/feature_management.rb
@@ -125,6 +125,16 @@ class FeatureManagement
       IdentityConfig.store.recaptcha_enterprise_project_id.present?
   end
 
+  # Whether we collect device profiling information as part of the account creation process
+  def self.account_creation_device_profiling_collecting_enabled?
+    case IdentityConfig.store.account_creation_device_profiling
+    when :enabled, :collect_only then true
+    when :disabled then false
+    else
+      raise 'Invalid value for account_creation_device_profiling'
+    end
+  end
+
   # Whether we collect device profiling information as part of the proofing process.
   def self.proofing_device_profiling_collecting_enabled?
     case IdentityConfig.store.proofing_device_profiling
